15.move inner class to upper level & move package: test -> main.
